+++ packages/core/src/builder.class.ts
@@ -276,12 +276,6 @@ type AllowEnrich =
 
 export type GetContentOptions = AllowEnrich & {
   /**
-   * Dictates which API endpoint is used when fetching content. Allows `'content'` and `'query'`.
-   * Defaults to `'query'`.
-   */
-  apiEndpoint?: 'content' | 'query';
-
-  /**
    * Optional fetch options to be passed as the second argument to the `fetch` function.
    */
   fetchOptions?: object;
@@ -996,13 +990,19 @@ export class Builder {
   }
 
   static isTrustedHost(hostname: string) {
-    return (
+    const isTrusted =
       this.trustedHosts.findIndex(trustedHost =>
         trustedHost.startsWith('*.')
           ? hostname.endsWith(trustedHost.slice(1))
           : trustedHost === hostname
-      ) > -1
-    );
+      ) > -1;
+
+    return isTrusted;
+  }
+
+  static isTrustedHostForEvent(event: MessageEvent) {
+    const url = parse(event.origin);
+    return url.hostname && Builder.isTrustedHost(url.hostname);
   }
 
   static runAction(action: Action | string) {
@@ -1864,11 +1864,8 @@ export class Builder {
   private bindMessageListeners() {
     if (isBrowser) {
       addEventListener('message', event => {
-        const url = parse(event.origin);
-        const isRestricted =
-          ['builder.register', 'builder.registerComponent'].indexOf(event.data?.type) === -1;
-        const isTrusted = url.hostname && Builder.isTrustedHost(url.hostname);
-        if (isRestricted && !isTrusted) {
+        const isTrusted = Builder.isTrustedHostForEvent(event);
+        if (!isTrusted) {
           return;
         }
 
@@ -2447,8 +2444,6 @@ export class Builder {
 
     const queue = useQueue || (usePastQueue ? this.priorContentQueue : this.getContentQueue) || [];
 
-    const apiEndpoint = queue[0].apiEndpoint || 'query';
-
     // TODO: do this on every request send?
     this.getOverridesFromQueryString();
 
@@ -2542,6 +2537,9 @@ export class Builder {
       }
     }
 
+    const isApiCallForCodegen =
+      queue[0].options?.format === 'solid' || queue[0].options?.format === 'react';
+
     for (const options of queue) {
       const format = options.format;
 
@@ -2580,7 +2578,7 @@ export class Builder {
       for (const key of properties) {
         const value = options[key];
         if (value !== undefined) {
-          if (apiEndpoint === 'query') {
+          if (isApiCallForCodegen) {
             queryParams.options = queryParams.options || {};
             queryParams.options[options.key!] = queryParams.options[options.key!] || {};
             queryParams.options[options.key!][key] = JSON.stringify(value);
@@ -2606,10 +2604,8 @@ export class Builder {
     }
 
     const format = queryParams.format;
-    const isApiCallForCodegen = format === 'solid' || format === 'react';
-    const isApiCallForCodegenOrQuery = isApiCallForCodegen || apiEndpoint === 'query';
 
-    if (apiEndpoint === 'content') {
+    if (!isApiCallForCodegen) {
       queryParams.enrich = true;
       if (queue[0].query) {
         const flattened = this.flattenMongoQuery({ query: queue[0].query });
@@ -2633,8 +2629,6 @@ export class Builder {
     let url;
     if (isApiCallForCodegen) {
       url = `${host}/api/v1/codegen/${this.apiKey}/${keyNames}`;
-    } else if (apiEndpoint === 'query') {
-      url = `${host}/api/v3/query/${this.apiKey}/${keyNames}`;
     } else {
       url = `${host}/api/v3/content/${queue[0].model}`;
     }
@@ -2660,7 +2654,7 @@ export class Builder {
             if (!observer) {
               return;
             }
-            const data = isApiCallForCodegenOrQuery ? result[keyName] : result.results;
+            const data = isApiCallForCodegen ? result[keyName] : result.results;
             const sorted = data; // sortBy(data, item => item.priority);
             if (data) {
               const testModifiedResults = Builder.isServer