+++ packages/react-tests/react-vite/src/App.tsx
@@ -1,5 +1,5 @@
-import { BuilderComponent, builder } from '@builder.io/react';
-import { getAPIKey, getProps } from '@sdk/tests';
+import { Builder, BuilderComponent, builder } from '@builder.io/react';
+import { getAPIKey, getProps, PAGES } from '@sdk/tests';
 import { useEffect, useState } from 'react';
 
 import '@builder.io/widgets';
@@ -34,6 +34,10 @@ function App() {
     builder.apiVersion = props?.apiVersion;
   }
 
+  if (props?.trustedHosts) {
+    Builder.trustedHosts = props.trustedHosts;
+  }
+
   // only enable tracking if we're not in the `/can-track-false` and `symbol-tracking` test route
   useEffect(() => {
     if (
@@ -44,14 +48,7 @@ function App() {
     }
   }, []);
 
-  /**
-   * - certain tests expect the content to only render after the first render
-   * - the `/large-reactive-state-editing` requires the `BuilderComponent` to
-   * be rendered immediately, so that the API request is made.
-   */
-  // issues with react types incompatibility (v16 vs v17 vs v18?)
-  // @ts-ignore
-  return props || window.location.pathname.includes('/large-reactive-state-editing') ? (
+  return props || PAGES[window.location.pathname]?.isGen1VisualEditingTest ? (
     <BuilderComponent {...props} />
   ) : (
     <div>Content Not Found</div>