+++ packages/sdks/e2e/angular-ssr/src/app/catch-all.component.ts
@@ -3,8 +3,7 @@ import { Component } from '@angular/core';
 // eslint-disable-next-line @typescript-eslint/consistent-type-imports
 import { ActivatedRoute } from '@angular/router';
 import type { RegisteredComponent } from '@builder.io/sdk-angular';
-import { BuilderBlockWithClassNameComponent } from './builder-block-with-class-name.component';
-import { HelloComponent } from './hello.component';
+import { customComponents } from './custom-components';
 
 interface BuilderProps {
   apiVersion: string;
@@ -14,6 +13,7 @@ interface BuilderProps {
   model: string;
   content: any;
   data?: any;
+  apiHost?: string;
 }
 
 @Component({
@@ -28,6 +28,7 @@ interface BuilderProps {
         [canTrack]="canTrack"
         [customComponents]="customComponents"
         [data]="data"
+        [apiHost]="apiHost"
       ></builder-content>
     </ng-container>
 
@@ -43,67 +44,9 @@ export class CatchAllComponent {
   model: BuilderProps['model'] = 'page';
   content: BuilderProps['content'];
   data: BuilderProps['data'];
+  apiHost: BuilderProps['apiHost'];
 
-  customComponents: RegisteredComponent[] = [
-    {
-      component: HelloComponent,
-      name: 'Hello',
-      inputs: [],
-      defaultChildren: [
-        {
-          '@type': '@builder.io/sdk:Element',
-          '@version': 2,
-          id: 'builder-ebca7d55d34f4fc9a6536600959cef5d',
-          component: {
-            name: 'Text',
-            options: {
-              text: 'inside an h1',
-            },
-          },
-        },
-      ],
-    },
-    {
-      name: 'BuilderBlockWithClassName',
-      component: BuilderBlockWithClassNameComponent,
-      shouldReceiveBuilderProps: {
-        builderBlock: true,
-        builderContext: true,
-        builderComponents: true,
-      },
-      inputs: [
-        {
-          name: 'content',
-          type: 'uiBlocks',
-          defaultValue: [
-            {
-              '@type': '@builder.io/sdk:Element',
-              '@version': 2,
-              id: 'builder-c6e179528dee4e62b337cf3f85d6496f',
-              component: {
-                name: 'Text',
-                options: {
-                  text: 'Enter some text...',
-                },
-              },
-              responsiveStyles: {
-                large: {
-                  display: 'flex',
-                  flexDirection: 'column',
-                  position: 'relative',
-                  flexShrink: '0',
-                  boxSizing: 'border-box',
-                  marginTop: '20px',
-                  lineHeight: 'normal',
-                  height: 'auto',
-                },
-              },
-            },
-          ],
-        },
-      ],
-    },
-  ];
+  customComponents: RegisteredComponent[] = customComponents;
 
   constructor(private activatedRoute: ActivatedRoute) {}
 
@@ -113,6 +56,7 @@ export class CatchAllComponent {
       this.canTrack = data.content?.canTrack;
       this.trustedHosts = data.content?.trustedHosts;
       this.data = data.content?.data;
+      this.apiHost = data.content?.apiHost;
       this.model = data.content?.model;
     });
   }