+++ packages/sdks/e2e/angular/src/app/app.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { Blocks, Content } from '@builder.io/sdk-angular';
 import { AppComponent } from './app.component';
-import { BuilderBlockWithClassNameComponent } from './builder-block-with-class-name.component';
+import { BuilderBlockWithClassNameComponent } from './custom-components/builder-block-with-class-name.component';
 
 @NgModule({
   declarations: [AppComponent, BuilderBlockWithClassNameComponent],