+++ packages/sdks/src/functions/evaluate/node-runtime/node-runtime.ts
@@ -86,8 +86,9 @@ let IVM_CONTEXT: Context | null = null;
  * `safeDynamicRequire` trick to import the `isolated-vm` package.
  */
 export const setIvm = (ivm: IsolatedVMImport, options: IsolateOptions = {}) => {
-  IVM_INSTANCE = ivm;
+  if (IVM_INSTANCE) return;
 
+  IVM_INSTANCE = ivm;
   setIsolateContext(options);
 };
 
@@ -121,6 +122,8 @@ const getIvm = (): IsolatedVMImport => {
 };
 
 function setIsolateContext(options: IsolateOptions = { memoryLimit: 128 }) {
+  if (IVM_CONTEXT) return IVM_CONTEXT;
+
   const ivm = getIvm();
   const isolate = new ivm.Isolate(options);
   const context = isolate.createContextSync();
@@ -143,11 +146,7 @@ function setIsolateContext(options: IsolateOptions = { memoryLimit: 128 }) {
 }
 
 const getIsolateContext = () => {
-  if (IVM_CONTEXT) return IVM_CONTEXT;
-
-  const context = setIsolateContext();
-
-  return context;
+  return setIsolateContext();
 };
 
 export const runInNode = ({