+++ packages/widgets/src/functions/async-component.tsx
@@ -56,5 +56,5 @@ const Loading = () => (
 //   })
 
 function reload(): any {
-  location.reload(true);
+  location.reload();
 }